requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
eth-account>=0.9.0
base58>=2.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
jinja2>=3.1.0
bittensor==9.12.2
bittensor-cli==9.16.0
substrate-interface
bcrypt
